Remove custom error msg on change
The custom error message was persisting on forms that don't watch the
field values, ie. most of them. It was intended that normalize_labels()
would be called before the form is actually submitted, but some
browsers ( at least chrome ) fire the checkValidity() before triggering
the submit event. This means the normalize_labels() is never called and
has no chance to clear the custom error message before checkValidity().